refacotr agent context to take in external components
